Pull up common code of aggregate mojos to parent
